random number of loot drops

+++ appserver/src/game/game.h
@@ -36,7 +36,7 @@ namespace Game {
 //    Units::Objects objects;
     std::array<std::unordered_map<Component::Position, Units::Objects>, 26> objects;
     //up to 4 items per tile
-    std::array<std::unordered_map<Component::Position, std::unordered_map<Component::Position, std::array<int , 4>>>, 26> items{};
+    std::array<std::unordered_map<Component::Position, std::unordered_map<Component::Position, std::array<uint8_t , 4>>>, 26> items{};
 
     //current turn
 //    int time = 0;
